oct-2122: working e2e tests
